+++ Makefile
@@ -3,6 +3,10 @@ NAME = minishell
 INCLUDES = includes/
 SOURCES = srcs/
 S_PARSE = parser/
+S_ERROR = error/
+S_EXEC = exec/
+S_BUILTIN = builtin/
+S_ENV = env/
 
 CC = cc
 # CFLAGS = -Wall -Wextra -Werror
@@ -10,12 +14,25 @@ AR = ar rcs
 RM = rm -rf
 
 PARSER += ms_tokenizer.c ms_tokenizer2.c ms_init.c ms_clear.c ms_lexer.c ms_unterminal.c ms_unterminal2.c ms_terminal.c ms_terminal2.c ms_parser.c 
+ERROR += ms_error_utils.c ms_error.c
+EXEC += ms_exec.c
+BUILTIN += ms_cd.c ms_echo.c ms_env.c ms_exit.c ms_export.c ms_pwd.c ms_unset.c
+ENV += ms_env_lst_utils.c ms_env_lst.c ms_env_serialize_utils.c ms_env_serialize.c
+
 
 M_FILES += minishell.c
 M_FILES += $(addprefix $(S_PARSE), $(PARSER))
+# M_FILES += $(addprefix $(S_ERROR), $(ERROR))
+# M_FILES += $(addprefix $(S_EXEC), $(EXEC))
+# M_FILES += $(addprefix $(S_BUILTIN), $(BUILTIN))
+# M_FILES += $(addprefix $(S_ENV), $(ENV))
 
 B_FILES += minishell.c
 B_FILES += $(addprefix $(S_PARSE), $(PARSER))
+# M_FILES += $(addprefix $(S_ERROR), $(ERROR))
+# M_FILES += $(addprefix $(S_EXEC), $(EXEC))
+# M_FILES += $(addprefix $(S_BUILTIN), $(BUILTIN))
+# M_FILES += $(addprefix $(S_ENV), $(ENV))
 
 H_FILES = \
 minishell.h
