+++ srcs/parser/ms_parser.c
@@ -15,7 +15,7 @@
 #include <unistd.h>
 #include <readline/readline.h>
 #include <readline/history.h>
-#include "minishell.h"
+#include "ms_parser.h"
 #include "../libft/includes/libft.h"
 
 int	ms_parser(t_ast *ast, t_token **token, size_t size)
