+++ srcs/parser/ms_tokenizer.c
@@ -6,7 +6,7 @@
 /*   By: jiwojung <jiwojung@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/25 18:29:36 by jiwojung          #+#    #+#             */
-/*   Updated: 2024/03/21 09:54:59 by jiwojung         ###   ########.fr       */
+/*   Updated: 2024/03/21 10:25:56 by jiwojung         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 #include <unistd.h>
 #include <readline/readline.h>
 #include <readline/history.h>
-#include "minishell.h"
+#include "ms_parser.h"
 #include "../libft/includes/libft.h"
 
 /**
